feat: Add new ProcessTable component in processes table
